Bugfix: error when deleting entities using service classes

+++ src/main/java/com/myproject/bikereviewapp/service/impl/BrandServiceImpl.java
@@ -95,7 +95,9 @@ public class BrandServiceImpl implements BrandService {
 
         log.debug("Removing Brand with ID: {}", id);
 
-        imageService.delete(brand.getImage().getId());
+        if(brand.getImage() != null) {
+            imageService.delete(brand.getImage().getId());
+        }
         brandRepository.delete(getById(id));
     }
